Switch to classy implementation
After adding inject, I started to feel like the namedtuple closure thing
wasn't super scalable. I mean, it might run faster, and be easier for me
to read without all the `self._<...>` noise, but I think in this case, a
class is the right tool for the job; managing a little blob of mutable
state.